Similarity percentage working

+++ DocCompare/src/ie/gmit/sw/Comparator.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface Comparator {
 	
 	// Compare method
-	public int compare(List<Integer>a, List<Integer>b); // Pass two sets (a and b) to the compare method
+	public float compare(List<Integer>a, List<Integer>b); // Pass two sets (a and b) to the compare method
 
 }
